Куча всего переделано

+++ ExchangeHub.FinanceService/Controllers/FinanceController.cs
@@ -23,7 +23,7 @@ public class FinanceController : ControllerBase
     [HttpGet("favorites")]
     public async Task<IActionResult> GetFavorites()
     {
-        var userId = long.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+        var userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
         var result = await _mediator.Send(new GetUserFavoriteCurrenciesQuery(userId));
         return Ok(result);
     }
